Workers/tests/submissions/Opera/semantics/navigation/002.html

<!--

/*
-->
<!doctype html>
<title>navigating 2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var date;
var newDate;
</script>
<iframe src="001-1.html"></iframe>
<script>
async_test(function() {
  var iframe = document.querySelector('iframe');
  iframe.src = "001-1.html"
  iframe.onload = this.step_func(function() {
    window[0].document.links[0].click();
    iframe.onload = this.step_func(function() {
      this.onload = null;
      newDate = new Date();
      setTimeout(this.step_func(function() {
        assert_greater_than(date, newDate);
        assert_equals(window[0].document.title, 'foo');
        this.done();
      }, 150);
    }
  });
});

</script>
<!--
*/
//-->

